new Sale Return, User activity Log, Module wise settings/Report, fixed some bugs,

+++ app/Http/Controllers/BankController.php
@@ -3,13 +3,16 @@
 namespace App\Http\Controllers;
 
 use App\Models\Bank;
+use App\Utils\UserActivityLogUtil;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Cache;
 
 class BankController extends Controller
 {
-    public function __construct()
+    protected $userActivityLogUtil;
+    public function __construct(UserActivityLogUtil $userActivityLogUtil)
     {
+        $this->userActivityLogUtil = $userActivityLogUtil;
         $this->middleware('auth:admin_and_user');
     }
     
@@ -17,8 +20,10 @@ class BankController extends Controller
     public function index()
     {
         if (auth()->user()->permission->accounting['ac_access'] == '0') {
+
             abort(403, 'Access Forbidden.');
         }
+
         return view('accounting.banks.index');
     }
 
@@ -37,11 +42,17 @@ class BankController extends Controller
             'branch_name' => 'required',
         ]);
 
-        Bank::insert([
+        $addBank = Bank::create([
             'name' => $request->name,
             'branch_name' => $request->branch_name,
             'address' => $request->address,
         ]);
+
+        $this->userActivityLogUtil->addLog(
+            action: 1,
+            subject_type: 16,
+            data_obj: $addBank
+        );
         
         return response()->json('Bank created successfully');
     }
@@ -60,14 +71,33 @@ class BankController extends Controller
             'branch_name' => $request->branch_name,
             'address' => $request->address,
         ]);
+
+        $this->userActivityLogUtil->addLog(
+            action: 2,
+            subject_type: 16,
+            data_obj: $updateBank
+        );
         
         return response()->json('Bank updated successfully');
     }
 
     public function delete(Request $request, $bankId)
     {
-        $deleteBank = Bank::find($bankId);
+        $deleteBank = Bank::with(['accounts'])->where('id', $bankId)->first();
+
         if (!is_null($deleteBank)) {
+
+            if(count($deleteBank->accounts) > 0) {
+
+                return response()->json('Can not be deleted, This bank has one or more account.');
+            }
+
+            $this->userActivityLogUtil->addLog(
+                action: 3,
+                subject_type: 16,
+                data_obj: $deleteBank
+            );
+
             $deleteBank->delete();  
         }
